+++ addons/l10n_et_reports/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from . import l10n_et_report_period_mixin
+from . import l10n_et_vat_declaration
+from . import l10n_et_wht_summary